refactor: inject experience repository
