+++ src/apps/server/experiences/providers/experience.provider.ts
@@ -2,12 +2,16 @@ import { AiResumeRepository } from '@libs/modules/database/repositories/aiResume
 import { AiResumeRepositoryImpl } from '@libs/modules/database/repositories/aiResume/aiResume.repository';
 import { CapabilityRepository } from '@libs/modules/database/repositories/capability/capability.interface';
 import { CapabilityRepositoryImpl } from '@libs/modules/database/repositories/capability/capability.repository';
-import { ExperienceRepository } from '@libs/modules/database/repositories/experience/experience.repository';
+import { ExperienceRepository } from '@libs/modules/database/repositories/experience/experience.interface';
+import { ExperienceRepositoryImpl } from '@libs/modules/database/repositories/experience/experience.repository';
 import { ExperienceCapabilityRepository } from '@libs/modules/database/repositories/experienceCapability.repository';
 import { ExperienceInfoRepository } from '@libs/modules/database/repositories/experienceInfo.repository';
 
 export const experienceProviders = [
-  ExperienceRepository,
+  {
+    provide: ExperienceRepository,
+    useClass: ExperienceRepositoryImpl,
+  },
   ExperienceInfoRepository,
   {
     provide: CapabilityRepository,
