+++ src/libs/modules/database/repositories/experience/experience.interface.ts
@@ -4,6 +4,10 @@ import { ExperienceSelect } from '@apps/server/experiences/interfaces/experience
 
 export interface ExperienceRepository {
   selectOneById(experienceId: number, select: ExperienceSelect): Promise<Partial<Experience & { ExperienceInfo?: ExperienceInfo }>>;
+  countByUserId(userId: number);
+  getExperienceById(experienceId: number);
+  getStarFromExperienceByExperienceId(experienceId: number);
+  findOneById(experienceid: number, userId: number);
   selectOneByUserId(
     userId: number,
     select: ExperienceSelect,
@@ -15,6 +19,10 @@ export interface ExperienceRepository {
     >
   >;
   findOneByUserId(userId: number): Promise<Experience>;
+  deleteById(experienceId: number);
+  getCount;
+  getExperiences;
+  getExperienceCardInfo;
 }
 
 export const ExperienceRepository = Symbol('ExperienceRepository');
